get order

+++ src/main/kotlin/com/wutsi/checkout/access/dto/OrderSummary.kt
@@ -2,7 +2,6 @@ package com.wutsi.checkout.access.dto
 
 import org.springframework.format.`annotation`.DateTimeFormat
 import java.time.OffsetDateTime
-import kotlin.Double
 import kotlin.Long
 import kotlin.String
 
@@ -12,8 +11,8 @@ public data class OrderSummary(
     public val storeId: Long = 0,
     public val status: String = "",
     public val subTotalPrice: Long = 0,
-    public val totalDiscount: Double = 0.0,
-    public val totalPrice: Double = 0.0,
+    public val totalDiscount: Long = 0,
+    public val totalPrice: Long = 0,
     public val currency: String = "",
     @get:DateTimeFormat(pattern = "yyyy-MM-dd'T'HH:mm:ssZ")
     public val created: OffsetDateTime = OffsetDateTime.now(),
